moved more tests to new bucket
