+++ dev/io.openliberty.microprofile.telemetry.1.0.internal.config_fat/fat/src/io/openliberty/microprofile/telemetry/internal/config_fat/FATSuite.java
@@ -28,8 +28,10 @@ import io.openliberty.microprofile.telemetry.internal_fat.shared.TelemetryAction
                 TelemetryConfigNullTest.class,
                 TelemetryConfigRuntimeModeIgnoresMPConfigTest.class,
                 TelemetryDisabledProvidersTest.class,
-                TelemetryMisconfigTest.class,
                 TelemetryDisabledTest.class,
+                TelemetryLoggingExporterTest.class,
+                TelemetryMisconfigTest.class,
+                TelemetrySpiTest.class,
 
 })
 public class FATSuite {
