+++ dev/io.openliberty.microprofile.telemetry.1.0.internal.config_fat/fat/src/io/openliberty/microprofile/telemetry/internal/config_fat/spi/sampler/SamplerTestServlet.java
@@ -10,7 +10,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package io.openliberty.microprofile.telemetry.internal_fat.apps.spi.sampler;
+package io.openliberty.microprofile.telemetry.internal.config_fat.spi.sampler;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
